Convert double to decimal type

+++ src/Fynance/Result/FyDividend.cs
@@ -2,10 +2,10 @@
 
 namespace Fynance.Result
 {
-    public class FyDividend
-    {
-        public DateTime Date { get; set; }
+	public class FyDividend
+	{
+		public DateTime Date { get; set; }
 
-        public double Value { get; set; }
-    }
+		public decimal Value { get; set; }
+	}
 }
