+++ src/Fynance/Yahoo/YSplit.cs
@@ -2,18 +2,18 @@
 
 namespace Fynance.Yahoo
 {
-    internal class YSplitResponse
-    {
-        [JsonProperty("date")] 
-        public double Date { get; set; }
+	internal class YSplitResponse
+	{
+		[JsonProperty("date")]
+		public double Date { get; set; }
 
-        [JsonProperty("numerator")]
-        public double Numerator { get; set; }
+		[JsonProperty("numerator")]
+		public decimal Numerator { get; set; }
 
-        [JsonProperty("denominator")] 
-        public double Denominator { get; set; }
+		[JsonProperty("denominator")]
+		public decimal Denominator { get; set; }
 
-        [JsonProperty("splitRatio")]
-        public string SplitRatio { get; set; }
-    }
+		[JsonProperty("splitRatio")]
+		public string SplitRatio { get; set; }
+	}
 }
